Event generation now working
Code now generates a list of events. Needs some tidying, and making it
read either a config file or take command line parameters would be a
good step.

Also, variable percentages for different products and ports would be a
good thing to add to make this flexible, and these should be represented
in the filename of the saved output file.

Working for now, though.