Improved status messaging and probability choice
Added messaging to show parameters and progress during event generation.

Created an example plot at end of file.

Changed event generation to use a negative binomial, but this still
needs a lot of tweaking to make it generate reasonable and interesting
time series.

Updated the term "species" to "product" throughout, to avoid confusion,
including in the name generation CSV file.